backend/requirements.txt
# --- Core Framework ---
fastapi>=0.110.0            # The main web framework
uvicorn[standard]>=0.29.0   # ASGI server to run FastAPI
pydantic>=2.7.0             # Data validation (strict typing)
pydantic-settings>=2.2.0    # Managing .env variables securely
email-validator>=2.1.0      # Email validation for Pydantic EmailStr

# --- Database & ORM ---
sqlalchemy>=2.0.29          # ORM for database interaction
alembic>=1.13.1             # Database migrations (schema changes)
asyncpg>=0.29.0             # High-speed Async PostgreSQL driver
psycopg2-binary>=2.9.9      # Standard driver (needed for Alembic synchronous ops)

# --- Authentication & Security ---
python-jose[cryptography]>=3.3.0  # JWT Token generation/decoding
passlib[bcrypt]>=1.7.4      # Password hashing
bcrypt>=4.0.0               # BCrypt backend (explicit)
python-multipart>=0.0.9     # Handling Form data (login requests)

# --- Real-Time & Async Tasks ---
redis>=5.0.3                # Redis client for Pub/Sub
websockets>=12.0            # For managing WebSocket connections to Frontend
httpx>=0.27.0               # Async HTTP client (for fetching external API data)

# --- Trading Logic (Jesse Engine Requirements) ---
numpy>=1.26.4               # Math calculations for indicators
pandas>=2.2.1               # DataFrame manipulation (OHLCV data)
scipy>=1.13.0               # Advanced scientific math
arrow>=1.3.0                # Better date/time handling for timezones
ccxt>=4.2.0                 # Universal crypto exchange API wrapper (critical)
ta-lib-bin                  # Technical Analysis library (RSI, MACD calculations)

# --- AWS Cloud Integration ---
boto3>=1.34.80              # AWS SDK (S3 upload, Secrets Manager)

# --- Testing ---
pytest>=8.1.1               # Testing framework
pytest-asyncio>=0.23.6      # Async support for tests
httpx>=0.27.0               # Test client for API endpoints

# --- Code Quality & Linting ---
ruff>=0.3.0                 # Fast Python linter
black>=24.3.0               # Code formatter
loguru>=0.7.2               # Structured logging

# --- Logging (retained from original) ---
loguru>=0.7.2               # Enhanced logging
python-dotenv>=1.0.0        # Environment variable loading
